+++ ogame/Technologies/EnergyTechnology.cpp
@@ -4,5 +4,5 @@
 EnergyTechnology::EnergyTechnology(Resources default_cost): Technology(default_cost)
 {
 	name = "Energy Technology";
-	m_requirements.push_back(std::make_pair(globals::Upgradables::LABORATORY, 1));
+	m_requirements.push_back(std::make_pair(static_cast<int>(globals::Upgradables::LABORATORY), 1));
 }
