+++ ogame/Technologies/EspionageTechnology.cpp
@@ -4,5 +4,5 @@
 EspionageTechnology::EspionageTechnology(Resources default_cost): Technology(default_cost)
 {
 	name = "Espionage Technology";
-	m_requirements.push_back(std::make_pair(globals::Upgradables::LABORATORY, 3));
+	m_requirements.push_back(std::make_pair(static_cast<int>(globals::Upgradables::LABORATORY), 3));
 }
