+++ ogame/Technologies/ImpulseDrive.cpp
@@ -4,8 +4,8 @@
 ImpulseDrive::ImpulseDrive(Resources default_cost): Technology(default_cost)
 {
 	name = "Impulse Drive";
-	m_requirements.push_back(std::make_pair(globals::Upgradables::LABORATORY, 2));
-	m_requirements.push_back(std::make_pair(globals::Upgradables::ENERGY_TECHNOLOGY, 1));
+	m_requirements.push_back(std::make_pair(static_cast<int>(globals::Upgradables::LABORATORY), 2));
+	m_requirements.push_back(std::make_pair(static_cast<int>(globals::Upgradables::ENERGY_TECHNOLOGY), 1));
 }
